+++ src/cases/var_wr_case.cpp
@@ -307,7 +307,7 @@ int e3sm_io_case::var_wr_case(e3sm_io_config &cfg,
             i++;
 
             /* write to D*.offsets, 2D array */
-            IPUT_VARA_NOADV(vars[i], MPI_INT, decom.disps[j])
+            IPUT_VARA_NOADV(vars[i], MPI_LONG_LONG, decom.disps[j])
             i++;
 
             /* write to D*.lengths, 2D array */
